feat(post): crade post

+++ src/modules/course/course.service.ts
@@ -1,10 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { type CreateCourseDto } from './schemas/create-course.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { Course } from './mongoose/course.schema';
 import { Model } from 'mongoose';
 import { ConfigService } from '@nestjs/config';
 import { AppConfig } from 'src/config/app.config';
+import { AppException } from 'src/common/errors/app-exception';
+import { ErrorCodes } from 'src/common/errors/error-codes';
 
 @Injectable()
 export class CourseService {
@@ -14,7 +16,7 @@ export class CourseService {
   ) {}
   async create(courseDto: CreateCourseDto) {
     const appName = this.configService.get('app', { infer: true });
-    console.log('🚀 ~ CourseService ~ create ~ appName:', appName?.env);
+
     const result = await this.courseModel.create(courseDto);
     return result;
   }
@@ -29,7 +31,17 @@ export class CourseService {
     return result;
   }
 
-  async update(id: string, courseDto: CreateCourseDto) {
+  async update(id: string, courseDto: Partial<CreateCourseDto>) {
+    console.log('🚀 ~ CourseService ~ update ~ courseDto:', courseDto);
+    const findExist = await this.courseModel.findById(id).exec();
+    if (!findExist) {
+      throw new AppException(
+        ErrorCodes.NOT_FOUND,
+        'Course not found',
+        HttpStatus.NOT_FOUND,
+        { courseId: id },
+      );
+    }
     const result = await this.courseModel
       .findByIdAndUpdate(id, courseDto, { new: true })
       .exec();
@@ -37,7 +49,16 @@ export class CourseService {
   }
 
   async delete(id: string) {
+    const findExist = await this.courseModel.findById(id).exec();
+    if (!findExist) {
+      throw new AppException(
+        ErrorCodes.NOT_FOUND,
+        'Course not found',
+        HttpStatus.NOT_FOUND,
+        { courseId: id },
+      );
+    }
     const result = await this.courseModel.findByIdAndDelete(id).exec();
-    return result;
+    return null;
   }
 }
